moved genome graph related functions and data to the GenomeGraph class and changed the type in classes effected by this change
